+++ docs/webpack.config.js
@@ -1,68 +1,85 @@
 const path = require("path");
 const webpack = require("webpack");
-const fableUtils = require("fable-utils");
 const HtmlWebpackPlugin = require('html-webpack-plugin');
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const UglifyJSPlugin = require('uglifyjs-webpack-plugin');
 
-function resolve(filePath) {
-    return path.join(__dirname, filePath)
-}
-
-var babelOptions = fableUtils.resolveBabelOptions({
-    presets: [["env", { "modules": false }]]
-});
+var babelOptions = {
+    presets: [
+        ["@babel/preset-env", {
+            "targets": {
+                "browsers": ["last 2 versions"]
+            },
+            "modules": false
+        }]
+    ]
+};
 
 var commonPlugins = [
     new HtmlWebpackPlugin({
-        filename: resolve('./public/index.html'),
-        template: resolve('./index.html'),
+        filename: './index.html',
+        template: './docs/index.html',
         hash: true,
         minify: isProduction ? {} : false
     })
 ];
 
 
-var isProduction = process.argv.indexOf("-p") >= 0;
+var isProduction = !process.argv.find(v => v.indexOf('webpack-dev-server') !== -1);
 console.log("Bundling for " + (isProduction ? "production" : "development") + "...");
 
 module.exports = {
     entry: isProduction ? // We don't use the same entry for dev and production, to make HMR over style quicker for dev env
         {
             demo: [
-                "babel-polyfill",
-                resolve('./docs.fsproj'),
-                resolve('./scss/main.scss')
+                "@babel/polyfill",
+                './docs/docs.fsproj',
+                './docs/scss/main.scss'
             ]
         } : {
             app: [
-                "babel-polyfill",
-                resolve('./docs.fsproj'),
+                "@babel/polyfill",
+                './docs/docs.fsproj',
             ],
             style: [
-                resolve('./scss/main.scss')
+                './docs/scss/main.scss'
             ]
         },
+    mode: isProduction ? "production" : "development",
     output: {
-        path: resolve('./public'),
+        path: path.join(__dirname, './public'),
         filename: isProduction ? '[name].[hash].js' : '[name].js'
+    },
+    optimization: {
+        // Split the code coming from npm packages into a different file.
+        // 3rd party dependencies change less often, let the browser cache them.
+        splitChunks: {
+            cacheGroups: {
+                commons: {
+                    test: /node_modules/,
+                    name: "vendors",
+                    chunks: "all"
+                }
+            }
+        },
     },
     plugins: isProduction ?
         commonPlugins.concat([
-            new ExtractTextPlugin('style.css'),
-            new webpack.optimize.CommonsChunkPlugin({
-                name: "manifest",
-                minChunks: Infinity
+            new MiniCssExtractPlugin({
+                filename: 'style.[hash].css'
+            }),
+            new UglifyJSPlugin({
+                uglifyOptions: {
+                    compress: { inline: false }
+                }
             })
         ])
         : commonPlugins.concat([
             new webpack.HotModuleReplacementPlugin(),
             new webpack.NamedModulesPlugin()
         ]),
-    resolve: {
-        modules: [resolve("../node_modules/")]
-    },
     devServer: {
-        contentBase: resolve('./public/'),
+        contentBase: './docs/public/',
         port: 8080,
         hot: true,
         inline: true
@@ -75,8 +92,7 @@ module.exports = {
                     loader: "fable-loader",
                     options: {
                         babel: babelOptions,
-                        define: isProduction ? [] : ["DEBUG"],
-                        plugins: path.join(__dirname, "./Plugins/bin/Release/netstandard2.0/Fable.Plugins.ImportExample.dll")
+                        define: isProduction ? [] : ["DEBUG"]
                     }
                 }
             },
@@ -89,8 +105,12 @@ module.exports = {
                 },
             },
             {
-                test: /\.s(a|c)ss$/,
-                use: ["style-loader", "css-loader", "sass-loader"]
+                test: /\.(sass|scss|css)$/,
+                use: [
+                    isProduction ? MiniCssExtractPlugin.loader : 'style-loader',
+                    'css-loader',
+                    'sass-loader',
+                ],
             },
             {
                 test: /\.css$/,